src/modules/cars/useCases/createCars/createCarsUseCase.ts
import { inject, injectable } from 'tsyringe';

import { Car } from '@modules/cars/infra/typeorm/entities/Car';
import { ICarRepository } from '@modules/cars/repositories/ICarsReposity';
import { AppError } from '@shared/errors/AppError';

interface IRequest {
  name: string;
  description: string;
  daily_rate: number;
  license_plate: string;
  fine_amount: number;
  brand: string;
  category_id: string;
}

@injectable()
class CreateCarsUseCase {
  constructor(
    @inject('CarsReposity')
    private carsRepository: ICarRepository,
  ) {}

  async execute({
    name,
    description,
    daily_rate,
    license_plate,
    fine_amount,
    brand,
    category_id,
  }: IRequest): Promise<Car> {
    const carAçreadyExists = await this.carsRepository.findByLicensePlate(
      license_plate,
    );

    if (carAçreadyExists) {
      throw new AppError('Car already exists');
    }

    const car = this.carsRepository.create({
      name,
      description,
      daily_rate,
      license_plate,
      fine_amount,
      brand,
      category_id,
    });

    return car;
  }
}

export { CreateCarsUseCase };
